top-tap-interfaces/requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
decorator==5.2.1
dogpile.cache==1.5.0
idna==3.11
iso8601==2.1.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
keystoneauth1==5.13.0
openstacksdk==4.10.0
os-service-types==1.8.2
pbr==7.0.3
platformdirs==4.9.2
psutil==7.2.2
pycparser==3.0
PyYAML==6.0.3
requests==2.32.5
requestsexceptions==1.4.0
setuptools==82.0.0
stevedore==5.6.0
typing_extensions==4.15.0
urllib3==2.6.3
